// src/app/raffles/raffles.interface.ts
export interface IRafflesQueryDTO {
    name: any;
    description?: any;
    price: any;
    total_numbers: any;
    status?: any;
    external_id?: any;
}

export interface IRafflesPersistDTO {
    name: any;
    description: any;
    price: any;
    total_numbers: any;
    status: any;
    external_id: any;
}